Continued progress with 2D Gaussian fitting
Appears to correctly fit a 2D ellipsoidal Gaussian to test data.  

Currently working to finalise this process for   real data - this involves correctly cropping the image to the current spot's area and transforming the centroid coordinates back.  A test sample indicates this isn't working quite right, as all sigma values are similar to the input, despite there being a range of actual PSF widths.