Measure object shape
Added simple module to count the number of voxels per object.  This will be extended as necessary at a later date.  It would be good to be able to generate 3D convex hulls for objects and use the measurements from the SegmentAnalysis plugin, but I'm not sure this is supported by Apache Commons Math.

Added measurement based filters to FilterObjects.

Implemented a wide JComboBox from StackOverflow.