Memory optimisations
Added RemoveImage module, which can completely remove an image from the workspace.  This is useful when the image is no longer used, as it frees up heap space memory.

Added "REMOVED_IMAGE" as a parameter, which is checked when populating the input image list.

Minor updates to modules to greater stability.

Noticed ObjectsToImage doesn't work for 4D objects yet (only the first frame is populated with objects).