Modified BatchProcessor to parallelise
Added option to BatchProcessor to run in parallel.  There are getter and setter options for this and for the number of threads (current default = 4).

This was shown to give an appropriate performance increase on a set of test images.

Following execution of an analysis, any images in the workspace have their pixel values set to null, which prevents a memory leak.