Updated clustering
Removed the previous "convex hull" method for generating the area of a cluster and moved to finite radius circles around each point in the circle.

Added method to RelateObjects to use children of an object.  In the present case these are the spots that make up a cluster.  These are used up to the same epsilon range as in DBSCAN.

Added SaveImages option to store output images in a directory structure mirroring the input structure.  This will allow us to store images on a separate HDD; hopefully increasing performance when IO-limited.