Added NormaliseIntensity module
Normalise intensity sets image intensity to minimum and maximum for that type (0-1 for 32-bit).  These are actually set as pixel values, not just for display.  This will hopefully be useful for things like RunTrackmate.

Standardised some verbose module output, so moduleName is accessible within each module.  It also automatically displays the relevant "Initialising" and "Completed" messages.