Minor bug fixes
File selectors should now not throw errors if no file is selected.

The correct number of submitted tasks should be shown when a batch process is started (previously was 1 less).

Added error to RidgeDetection module if the plugin fails to complete (very rare).  This should appear in the error stream System.err.