Coordinates now stored in .csv files
Coordinates for expected objects are now stored in csv files rather than hard-coded into methods.

Resolved the remaining issues preventing the code compiling.

Still fails on relating objects by surface; however, this is to be expected as the system doesn't work correctly when the objects are inside the parent object (should be negative distance).