Added GUI separator
Created a fake Parameter, which tells the basic GUI to add a JSeparator.  These can be added to split the parameters up and make the interface a bit easier to see.

Added a bug fix for the evaluation state of the modules as they are moved up and down.

Make some minor alterations to the module checkboxes.