Distance map can now use real spatial calibrationo
Previously distance map ignored the differences in XY and Z spatial calibration.  There is now an option to use this ratio.  When in use, prior to calculation of the distance map, the input image is interpoalted.  The reverse process is applied after the calculation.

Started adding tests for image math module.