Added specified calibration
Spatial units are now specified at the InputOptions stage.  As long as BioFormats is capable of reading the file and interpreting the units it can convert them to the specified units.  This means results column headers can state the correct units (rather than "CAL").

Switched ImageSequence loader to use BioFormats for compatibility with the above-stated functionality.