Changed module enabled button
Module enabled checkbox has now been changed to a button, which shows the universal symbol for standby.  This changes from green to red when the module is disabled.  It will hopefully be a little more intuitive.

The next thing to add to the GUI will be a general "view" button, similar to Cellprofiler, which provides a general control for displaying images.