Updated FocusStack
FocusStack can now work from a previously-calculated height map.  This means the same height map can be applied to different channels of the same image (each channel would need a separate instance of the Module).

Fixed a bug in ProjectImage where it projected along the channel if no stack was present.