Development of radial intensity profiles
Made distance-based intensity profiles more generic as well as adding an equivalent functionality to images (as well as objects).

Started adding saving and plotting intensity profiles.  This is currently on an individual file basis (as is image saving), but in the future it would be good to group them all together.