Added track merging module
MergeTracks module allows the user to specify a comma-separated list of tracks to combine.  The first track ID in the list is considered dominant, so will provide the track ID and the children in any timepoint disputes (currently tracks can only have a single child at a specific timepoint - branching may be an addition in the future).

Moved comma-separated string interpreting to a separate class.