Bugfix for RelateObjects
When no children were detected it was possible to get a divide by zero error from the writeMessage method.

Also, there appears to be an inconsistency in how ResolveObjectCoOccurrance works.  Presumably due to use of non-ordered maps/sets the order of evaluation results in slightly different results for dense samples.  This will be looked into over the next couple of days.